+++ packages/zod/src/v4/core/schemas.ts
@@ -168,9 +168,7 @@ export interface $ZodType<
   "~standard": $ZodStandardSchema<this>;
 }
 export interface _$ZodType<T extends $ZodTypeInternals = $ZodTypeInternals>
-  extends $ZodType<T["output"], T["input"], T> {
-  // _zod: T;
-}
+  extends $ZodType<T["output"], T["input"], T> {}
 
 export const $ZodType: core.$constructor<$ZodType> = /*@__PURE__*/ core.$constructor("$ZodType", (inst, def) => {
   inst ??= {} as any;
@@ -178,14 +176,6 @@ export const $ZodType: core.$constructor<$ZodType> = /*@__PURE__*/ core.$constru
   inst._zod.def = def; // set _def property
   inst._zod.bag = inst._zod.bag || {}; // initialize _bag object
   inst._zod.version = version;
-
-  Object.defineProperty(inst, "type", {
-    get() {
-      return this._zod.def.type;
-    },
-    enumerable: true,
-    configurable: true,
-  });
 
   const checks = [...(inst._zod.def.checks ?? [])];
